+++ src/screens/App_EditActivityScreen.tsx
@@ -73,7 +73,7 @@ const EditActivityScreen: React.FC<Props> = ({navigation, route}) => {
     const data = {
       activity_id: selected.id,
       date: moment(selectedDate).format('DD/MM/YYYY'),
-      count: parseInt(distanceTimeData),
+      count: parseFloat(distanceTimeData),
       is_distance: defaultOption === 0,
       calories: calminsteps.cal,
       min: calminsteps.min,
